feat(console): create orbiter args
