+++ src/console/src/api/factory.rs
@@ -4,7 +4,7 @@ use candid::Principal;
 use ic_cdk_macros::update;
 use junobuild_shared::ic::api::caller;
 use junobuild_shared::ic::UnwrapOrTrap;
-use junobuild_shared::types::interface::{CreateCanisterArgs, CreateSatelliteArgs};
+use junobuild_shared::types::interface::{CreateOrbiterArgs, CreateSatelliteArgs};
 
 #[update]
 async fn create_satellite(args: CreateSatelliteArgs) -> Principal {
@@ -16,7 +16,7 @@ async fn create_satellite(args: CreateSatelliteArgs) -> Principal {
 }
 
 #[update]
-async fn create_orbiter(args: CreateCanisterArgs) -> Principal {
+async fn create_orbiter(args: CreateOrbiterArgs) -> Principal {
     let caller = caller();
 
     create_orbiter_console(caller, args).await.unwrap_or_trap()
