+++ src/console/src/factory/mod.rs
@@ -1,6 +1,8 @@
 mod canister;
+mod impls;
 pub mod mission_control;
 pub mod orbiter;
 pub mod satellite;
 mod services;
+mod types;
 mod utils;
